society_backend/society_backend/requirements.txt
@"
fastapi==0.115.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.10.6
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.4.1
python-multipart==0.0.9
alembic==1.13.1
email-validator==2.1.1
"@ | Out-File -FilePath requirements.txt -Encoding utf8